Extracted functions relevanat to each model into their own files to clear up main.go
