+++ entity.go
@@ -1,0 +1,9 @@
+package main
+
+func (e *Entity) SetDescription(description string) {
+	e.Description = description
+}
+
+func (e *Entity) GetDescription() string {
+	return e.Description
+}
