+++ item.go
@@ -1,0 +1,9 @@
+package main
+
+func (i *Item) SetDescription(description string) {
+	i.Description = description
+}
+
+func (i *Item) GetDescription() string {
+	return i.Description
+}
