+++ main.go
@@ -9,255 +9,18 @@ import (
 	"strings"
 )
 
-func clearScreen() {
-    var cmd *exec.Cmd
-    if runtime.GOOS == "windows" {
-        cmd = exec.Command("cmd", "/c", "cls")
-    } else {
-        cmd = exec.Command("clear")
-    }
-    cmd.Stdout = os.Stdout
-    cmd.Run()
-}
-
-func (i *Item) SetDescription(description string) {
-    i.Description = description
-}
-
-func (i *Item) GetDescription() string {
-    return i.Description
-}
-
-func (r *Room) SetDescription(description string) {
-    r.Description = description
-}
-
-func (r *Room) GetDescription() string {
-    return r.Description
-}
-
 var validInteractions = []*Interaction{}
-
-func (e *Entity) SetDescription(description string) {
-    e.Description = description
-}
-
-func (e *Entity) GetDescription() string {
-    return e.Description
-}
-
-func (p *Player) Move(direction string) {
-	if p.CurrentEntity != nil {
-		p.CurrentEntity = nil
-	}
-	if newRoom, ok := p.CurrentRoom.Exits[direction]; ok {
-		p.CurrentRoom = newRoom
-
-		fmt.Printf("You are in %s\n", p.CurrentRoom.Name)
-	} else {
-		fmt.Println("You can't go that way!")
-	}
-}
-
 var plateOrder = []string{"first-plate", "second-plate", "third-plate", "fourth-plate", "fifth-plate", "sixth-plate"}
 var currentPlateIndex = 0
 var gameOver = false
 
-
-func (p *Player) Take(itemName string) {
-	item, ok := p.CurrentRoom.Items[itemName]
-	switch {
-	case !ok || item.Hidden:
-		fmt.Printf("You can't take %s\n", itemName)
-		return
-	case p.AvailableWeight < item.Weight:
-		fmt.Println("Weight limit reached! Please drop an item before taking more.")
-		return
-	case isPlate(itemName):
-		if itemName == plateOrder[currentPlateIndex] {
-			p.Inventory[item.Name] = item
-			p.ChangeCarriedWeight(item, "increase")
-			delete(p.CurrentRoom.Items, item.Name)
-			currentPlateIndex++
-
-			fmt.Printf("%s has been added to your inventory.\n", item.Name)
-		} else {
-			fmt.Println("As you attempt to grab the greasy plates without removing the ones stacked above them, they slip from your grasp and shatter, creating a chaotic mess.\n\nNow Rosie is very grumpy.")
-			gameOver = true
-		}
-		
-	default:
-		p.Inventory[item.Name] = item
-		p.ChangeCarriedWeight(item, "increase")
-		delete(p.CurrentRoom.Items, item.Name)
-
-		fmt.Printf("%s has been added to your inventory.\n", item.Name)
-	}
-}
-
 func isPlate(itemName string) bool {
-    for _, plate := range plateOrder {
-        if itemName == plate {
-            return true
-        }
-    }
-    return false
-}
-
-func (p *Player) ChangeCarriedWeight(item *Item, operation string) {
-	switch {
-	case operation == "increase":
-		p.CarriedWeight += item.Weight
-		p.AvailableWeight -= item.Weight
-		return
-	case operation == "decrease":
-		p.CarriedWeight -= item.Weight
-		p.AvailableWeight += item.Weight
-		return
-	}
-}
-
-func (p *Player) Drop(itemName string) {
-	if item, ok := p.Inventory[itemName]; ok {
-		if isPlate(itemName) {
-			println("You can't just leave those plates lying around! It's time to load them into the dishwasher!")
-			return
-		}
-
-		delete(p.Inventory, item.Name)
-		p.ChangeCarriedWeight(item, "decrease")
-		p.CurrentRoom.Items[item.Name] = item
-
-		fmt.Printf("You dropped %s.\n", item.Name)
-	} else {
-		fmt.Printf("You don't have %s.\n", itemName)
-	}
-}
-
-func (p *Player) ShowInventory() {
-	if len(p.Inventory) == 0 {
-		fmt.Printf("Your inventory is empty.\nAvailable space: %d\n", p.AvailableWeight)
-		return
-	}
-	fmt.Printf("Available space: %d\nYour inventory contains:\n", p.AvailableWeight)
-	for itemName, item := range p.Inventory {
-		fmt.Printf("- %s: %s Weight: %d\n", itemName, item.Description, item.Weight)
-	}
-}
-
-func (p *Player) ShowRoom() {
-    fmt.Printf("You are in %s\n\n%s\n", p.CurrentRoom.Name, p.CurrentRoom.Description)
-
-	if p.EntitiesArePresent() {
-			fmt.Println("\nYou can approach:")
-			for _, entity := range p.CurrentRoom.Entities {
-				switch {
-				case p.CurrentEntity != nil:
-					if entity.Name == p.CurrentEntity.Name {
-						fmt.Printf("- %s (currently approached)\n", entity.Name)
-					} else if !entity.Hidden{
-						fmt.Printf("- %s\n", entity.Name)
-					}
-				default:
-					if !entity.Hidden{
-						fmt.Printf("- %s\n", entity.Name)
-					}
-				}
-			}
-		}
-		
-	if p.ItemsArePresent() {
-			fmt.Println("\nThe room contains:")
-			for itemName, item := range p.CurrentRoom.Items {
-				if !item.Hidden {
-					fmt.Printf("- %s: %s Weight: %d\n", itemName, item.Description, item.Weight)
-				}
-			}
-		}
-	}
-
-	
-
-func (p *Player) ItemsArePresent() bool {
-	if len(p.CurrentRoom.Items) != 0 {
-	for _, item := range p.CurrentRoom.Items {
-		if !item.Hidden {
+	for _, plate := range plateOrder {
+		if itemName == plate {
 			return true
 		}
 	}
-}
 	return false
-}
-
-func (p *Player) EntitiesArePresent() bool {
-	if len(p.CurrentRoom.Entities) != 0 {
-	for _, entity := range p.CurrentRoom.Entities {
-		if !entity.Hidden {
-			return true
-		}
-	}
-}
-	return false
-}
-
-
-func (p *Player) Approach(entityName string) {
-	if p.CurrentEntity != nil {
-		p.CurrentEntity = nil
-	}
-	if entity, ok := p.CurrentRoom.Entities[entityName]; ok && !entity.Hidden{
-
-		p.CurrentEntity = entity
-		fmt.Println(entity.Description)
-	} else {
-		fmt.Printf("You can't approach %s.\n", entityName)
-	}
-}
-
-func (p *Player) Leave() {
-	if p.CurrentEntity != nil {
-		p.CurrentEntity = nil
-		p.ShowRoom()
-	} else {
-		fmt.Println("You have not approached anything. If you wish to leave the game, use the exit command.")
-	}
-}
-
-func (p *Player) ShowMap() {
-	for direction, exit := range p.CurrentRoom.Exits {
-		fmt.Printf("%s: %s\n", direction, exit.Name)
-	}
-}
-
-func (p *Player) Use(itemName string, target string) {
-	if p.CurrentEntity == nil {
-		fmt.Println("Approach to use an item.")
-		return
-	}	
-	if p.CurrentEntity.Name == target {
-		if _, ok := p.Inventory[itemName]; ok {
-				for _, interaction := range validInteractions {
-					if interaction.ItemName == itemName && interaction.EntityName == target {
-						p.TriggerEvent(interaction.Event)
-						p.ChangeCarriedWeight(p.Inventory[itemName], "decrease")
-						delete(p.Inventory, itemName)
-						return
-					}
-				}
-		} else {
-			fmt.Printf("You don't have %s.\n", itemName)
-			return
-		}
-	} else {
-		fmt.Printf("%s not found.\n", target)
-		return
-	}
-	fmt.Printf("You can't use %s on %s.\n", itemName, target)
-}
-
-func (p *Player) TriggerEvent(event *Event) {
-	fmt.Println(event.Outcome)
-	event.Triggered = true
 }
 
 func updateDescription(d Describable, newDescription string) {
@@ -268,10 +31,21 @@ func showCommands() {
 	fmt.Println("-exit -> quits the game\n\n-commands -> shows the commands\n\n-look -> shows the content of the room.\n\n-approach <entity> -> to approach an entity\n\n-leave -> to leave an entity\n\n-inventory -> shows items in the inventory\n\n-take <item> -> to take an item into your inventory\n\n-drop <item> -> to drop an item from your inventory and move it to the current room\n\n-use <item> -> to make use of a certain item when you approach an entity\n\n-move <direction> -> to move to a different room\n\n-map -> shows the directions you can take")
 }
 
+func clearScreen() {
+	var cmd *exec.Cmd
+	if runtime.GOOS == "windows" {
+		cmd = exec.Command("cmd", "/c", "cls")
+	} else {
+		cmd = exec.Command("clear")
+	}
+	cmd.Stdout = os.Stdout
+	cmd.Run()
+}
+
 func main() {
 	introduction := "It's the last day at the Academy, and you and your fellow graduates are ready to take on the final hack-day challenge.\nHowever, this time, it's different. Alan and Dan, your instructors, have prepared something more intense than ever before — a true test of your problem-solving and coding skills.\nThe doors to the academy are locked, the windows sealed. The only way out is to find and solve a series of riddles that lead to the terminal in a hidden room.\nThe challenge? Crack the code on the terminal to unlock the doors. But it's not that simple.\nYou'll need to gather items, approach Alan and Dan for cryptic tips, and outsmart the obstacles they've laid out for you.\nAs the tension rises, only your wits, teamwork, and knowledge can guide you to freedom.\nAre you ready to escape?\nOh and remember... You don't want to make Rosie grumpy! So don't do anything crazy.\n\nif at any point you feel lost, type 'commands' to display the list of all commands.\nThe command 'look' is always useful to get your bearings and see the options available to you.\nThe command 'exit' will make you quit the game at any time. Make sure you do mean to use it, or you will inadvertently lose all of your progress!"
 
-	introductionShown:= false
+	introductionShown := false
 
 	validInteractions = []*Interaction{
 		{
@@ -280,34 +54,34 @@ func main() {
 			Event:      &Event{Description: "get-your-lanyard", Outcome: "Cheers! I needed that... by the way, where is your lanyard? I must have forgotten to give it to you.\nYou'll need that to move between rooms, here it is.\n\n(lanyard can now be found in the room).\n", Triggered: false},
 		},
 		{
-			ItemName: "first-plate",
+			ItemName:   "first-plate",
 			EntityName: "dishwasher",
-			Event: &Event{Description: "first-plate-loaded", Outcome: "You loaded the first plate into the dishwasher.", Triggered: false},
+			Event:      &Event{Description: "first-plate-loaded", Outcome: "You loaded the first plate into the dishwasher.", Triggered: false},
 		},
 		{
-			ItemName: "second-plate",
+			ItemName:   "second-plate",
 			EntityName: "dishwasher",
-			Event: &Event{Description: "second-plate-loaded", Outcome: "You loaded the second plate into the dishwasher.", Triggered: false},
+			Event:      &Event{Description: "second-plate-loaded", Outcome: "You loaded the second plate into the dishwasher.", Triggered: false},
 		},
 		{
-			ItemName: "third-plate",
+			ItemName:   "third-plate",
 			EntityName: "dishwasher",
-			Event: &Event{Description: "third-plate-loaded", Outcome: "You loaded the third plate into the dishwasher.", Triggered: false},
+			Event:      &Event{Description: "third-plate-loaded", Outcome: "You loaded the third plate into the dishwasher.", Triggered: false},
 		},
 		{
-			ItemName: "fourth-plate",
+			ItemName:   "fourth-plate",
 			EntityName: "dishwasher",
-			Event: &Event{Description: "fourth-plate-loaded", Outcome: "You loaded the fourth plate into the dishwasher.", Triggered: false},
+			Event:      &Event{Description: "fourth-plate-loaded", Outcome: "You loaded the fourth plate into the dishwasher.", Triggered: false},
 		},
 		{
-			ItemName: "fifth-plate",
+			ItemName:   "fifth-plate",
 			EntityName: "dishwasher",
-			Event: &Event{Description: "fifth-plate-loaded", Outcome: "You loaded the fifth plate into the dishwasher.", Triggered: false},
+			Event:      &Event{Description: "fifth-plate-loaded", Outcome: "You loaded the fifth plate into the dishwasher.", Triggered: false},
 		},
 		{
-			ItemName: "sixth-plate",
+			ItemName:   "sixth-plate",
 			EntityName: "dishwasher",
-			Event: &Event{Description: "sixth-plate-loaded", Outcome: "You loaded the sixth plate into the dishwasher.", Triggered: false},
+			Event:      &Event{Description: "sixth-plate-loaded", Outcome: "You loaded the sixth plate into the dishwasher.", Triggered: false},
 		},
 	}
 
@@ -324,25 +98,25 @@ func main() {
 	staffRoom := Room{
 		Name:        "break-room",
 		Description: "A cozy lounge designed for both academy students and tutors, offering a welcoming space to unwind and socialise.\nComfortable seating invites you to relax, while the warm ambiance encourages lively conversations and friendly exchanges.",
-		Items:      make(map[string]*Item),
-		Entities:   make(map[string]*Entity),
-		Exits:      make(map[string]*Room),
+		Items:       make(map[string]*Item),
+		Entities:    make(map[string]*Entity),
+		Exits:       make(map[string]*Room),
 	}
 
 	codingLab := Room{
 		Name:        "coding-lab",
 		Description: "A bright, tech-filled room with sleek workstations, whiteboards, and collaborative spaces.\nThe air buzzes with creativity as students code, share ideas, and tackle challenges together.",
-		Items:      make(map[string]*Item),
-		Entities:   make(map[string]*Entity),
-		Exits:      make(map[string]*Room),
+		Items:       make(map[string]*Item),
+		Entities:    make(map[string]*Entity),
+		Exits:       make(map[string]*Room),
 	}
 
 	terminalRoom := Room{
 		Name:        "terminal-room",
 		Description: "As you step into the terminal room, you're greeted by the soft hum of machines and the flickering glow of monitors lining the walls.\n\nThe air is charged with a sense of urgency, filled with the scent of freshly brewed coffee mingling with the faint odor of electrical components.\n\nIn the center of the room, a sleek, state-of-the-art terminal stands atop a polished wooden desk.",
-		Items:      make(map[string]*Item),
-		Entities:   make(map[string]*Entity),
-		Exits:      make(map[string]*Room),
+		Items:       make(map[string]*Item),
+		Entities:    make(map[string]*Entity),
+		Exits:       make(map[string]*Room),
 	}
 
 	staffRoom.Exits["south"] = &codingLab
@@ -418,7 +192,7 @@ func main() {
 
 		if player.CurrentEntity != nil && player.CurrentEntity.Name == "kettle" {
 			tea.Hidden = false
-			kettle.SetDescription("A kettle — essential for survival, impossible to function without one nearby.");
+			kettle.SetDescription("A kettle — essential for survival, impossible to function without one nearby.")
 		}
 
 		if player.CurrentEntity != nil && player.CurrentEntity.Name == "desk" {
@@ -473,7 +247,6 @@ func main() {
 
 		fmt.Print("Enter command: ")
 
-
 		if scanner.Scan() {
 			input := scanner.Text()
 			input = strings.TrimSpace(input)
@@ -485,10 +258,8 @@ func main() {
 				break
 			}
 
-			
-
 			if isAttemptingPassword {
-				if remainingPasswordAttempts == 1 && input != computerPassword{
+				if remainingPasswordAttempts == 1 && input != computerPassword {
 					clearScreen()
 					fmt.Println("Alan's computer is locked, halting your progress in the challenge. To top it off, you've made Rosie grumpy, as she'll now have to take the computer to IT.\n\nThank you for playing!")
 					break
@@ -503,7 +274,7 @@ func main() {
 					dishwasher.Hidden = false
 				} else if input == "leave" {
 					isAttemptingPassword = false
-					} else {
+				} else {
 					remainingPasswordAttempts--
 					clearScreen()
 					fmt.Printf("Incorrect password. Try again, or type 'leave' to stop entering the password.\n\nRemaining attempts: %d\n\n", remainingPasswordAttempts)
@@ -519,7 +290,7 @@ func main() {
 					player.Leave()
 					continue
 				}
-			
+
 				if !IsFirstCommand {
 					if input == "cd /secret-files" {
 						clearScreen()
@@ -545,8 +316,6 @@ func main() {
 					}
 				}
 			}
-			
-			
 
 			parts := strings.Fields(input)
 			if len(parts) == 0 {
@@ -588,12 +357,12 @@ func main() {
 					if !unlockComputer.Triggered {
 						if player.CurrentEntity != nil && player.CurrentEntity.Name == "computer" {
 							isAttemptingPassword = true
+						}
 					}
-                }
-				if player.CurrentEntity != nil && player.CurrentEntity.Name == "terminal" {
-					isAttemptingTerminal = true
-				}
-                
+					if player.CurrentEntity != nil && player.CurrentEntity.Name == "terminal" {
+						isAttemptingTerminal = true
+					}
+
 				} else {
 					fmt.Println("Specify an entity to approach.")
 				}
@@ -628,10 +397,9 @@ func main() {
 			case computerPassword:
 				continue
 			default:
-					clearScreen()
-					fmt.Println("Unknown command:", command)
+				clearScreen()
+				fmt.Println("Unknown command:", command)
 			}
 		}
 	}
 }
-
