+++ room.go
@@ -1,0 +1,9 @@
+package main
+
+func (r *Room) SetDescription(description string) {
+	r.Description = description
+}
+
+func (r *Room) GetDescription() string {
+	return r.Description
+}
